// ui/src/features/sql/api/sql-schema.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type SqlSchemaQueryVariables = Exact<{
  query: string
}>

export type SqlSchemaQuery = {
  readonly sql: {
    readonly columns: ReadonlyArray<string>
    readonly rows: ReadonlyArray<string>
    readonly rowCount: number
    readonly truncated: boolean
  }
}

export const SqlSchemaDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "SqlSchema" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "query" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "sql" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "query" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "query" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "columns" } },
                { kind: "Field", name: { kind: "Name", value: "rows" } },
                { kind: "Field", name: { kind: "Name", value: "rowCount" } },
                { kind: "Field", name: { kind: "Name", value: "truncated" } },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<SqlSchemaQuery, SqlSchemaQueryVariables>
// ...
export const SqlSchemaQuerySchema: z.ZodType<SqlSchemaQuery> = z.object({
  sql: z.object({
    columns: z.array(z.string()),
    rows: z.array(z.string()),
    rowCount: z.number(),
    truncated: z.boolean(),
  }),
})