// ui/src/features/sql/api/sql-snippets-list.generated.ts
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
// ...
export type SqlSnippetsListQueryVariables = Exact<{
  page?: string | null | undefined
}>

export type SqlSnippetsListQuery = {
  readonly savedViews: ReadonlyArray<{
    readonly id: string
    readonly name: string
    readonly page: string
    readonly state: string
    readonly updatedAtNanos: string
  }>
}

export const SqlSnippetsListDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "query",
      name: { kind: "Name", value: "SqlSnippetsList" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: { kind: "Variable", name: { kind: "Name", value: "page" } },
          type: { kind: "NamedType", name: { kind: "Name", value: "String" } },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "savedViews" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "page" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "page" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                { kind: "Field", name: { kind: "Name", value: "id" } },
                { kind: "Field", name: { kind: "Name", value: "name" } },
                { kind: "Field", name: { kind: "Name", value: "page" } },
                { kind: "Field", name: { kind: "Name", value: "state" } },
                {
                  kind: "Field",
                  name: { kind: "Name", value: "updatedAtNanos" },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<
// ...
export const SqlSnippetsListQuerySchema: z.ZodType<SqlSnippetsListQuery> =
  z.object({
    savedViews: z.array(
      z.object({
        id: z.string(),
        name: z.string(),
        page: z.string(),
        state: z.string(),
        updatedAtNanos: z.string(),
      })
    ),
  })